+++ src/doomstat.h
@@ -92,6 +92,7 @@ extern char             speciallumpname[6];
 extern dboolean         freeze;
 
 extern dboolean         sigil;
+extern dboolean         buckethead;
 extern dboolean         nerve;
 extern dboolean         bfgedition;